+++ packages/error/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.18-beta.2](https://github.com/kamontat/kcutils/compare/@kcutils/error@0.4.18-beta.1...@kcutils/error@0.4.18-beta.2) (2020-09-09)
+
+**Note:** Version bump only for package @kcutils/error
+
+
+
+
+
+## [0.4.18-beta.1](https://github.com/kamontat/kcutils/compare/@kcutils/error@0.4.18-beta.0...@kcutils/error@0.4.18-beta.1) (2020-09-09)
+
+**Note:** Version bump only for package @kcutils/error
+
+
+
+
+
+## [0.4.18-beta.0](https://github.com/kamontat/kcutils/compare/@kcutils/error@0.4.17...@kcutils/error@0.4.18-beta.0) (2020-09-09)
+
+**Note:** Version bump only for package @kcutils/error
+
+
+
+
+
 ## [0.4.17](https://github.com/kamontat/kcutils/compare/@kcutils/error@0.4.16...@kcutils/error@0.4.17) (2020-09-05)
 
 **Note:** Version bump only for package @kcutils/error